+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- New `group_by` processor.
+
 ## 0.34.8 - 2018-10-10
 
 ### Added